Cleanup for submitting a PR
